+ cache is now filled even if the screen orientation changes + cache feeder is now a service + EntityOptimizer not attached to cache feeder (and should be renamed to something more accurate); it is now used by the activity to trigger a cache update when the db is updated. - ProjectsRepository does not call the cache feeder and only shows data
