+++ app/src/main/AndroidManifest.xml
@@ -29,5 +29,6 @@
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
+        <service android:name=".cache.optimizer.CacheFeederService" android:exported="false"/>
     </application>
 </manifest>
