+++ app/src/main/java/org/rares/miner49er/_abstract/ResizeableItemsUiOps.java
@@ -101,7 +101,7 @@ public abstract class ResizeableItemsUiOps
 
         setupSelectedItemAnimation(holder);
 
-        int adapterPosition = holder.getAdapterPosition();
+        int adapterPosition = holder.getAdapterPosition();  /// if holder ==null ???
         boolean enlarge = selectItem(adapterPosition);
 
         domainLink.onParentSelected(holder.getItemProperties(), enlarge);
