+++ app/src/main/java/org/rares/miner49er/domain/issues/persistence/IssuesRepository.java
@@ -16,19 +16,6 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-import static org.rares.miner49er.cache.Cache.CACHE_EVENT_REMOVE_ENTRY;
-import static org.rares.miner49er.cache.Cache.CACHE_EVENT_REMOVE_ISSUE;
-import static org.rares.miner49er.cache.Cache.CACHE_EVENT_REMOVE_PROJECT;
-import static org.rares.miner49er.cache.Cache.CACHE_EVENT_REMOVE_USER;
-import static org.rares.miner49er.cache.Cache.CACHE_EVENT_UPDATE_ENTRIES;
-import static org.rares.miner49er.cache.Cache.CACHE_EVENT_UPDATE_ENTRY;
-import static org.rares.miner49er.cache.Cache.CACHE_EVENT_UPDATE_ISSUE;
-import static org.rares.miner49er.cache.Cache.CACHE_EVENT_UPDATE_ISSUES;
-import static org.rares.miner49er.cache.Cache.CACHE_EVENT_UPDATE_PROJECT;
-import static org.rares.miner49er.cache.Cache.CACHE_EVENT_UPDATE_PROJECTS;
-import static org.rares.miner49er.cache.Cache.CACHE_EVENT_UPDATE_USER;
-import static org.rares.miner49er.cache.Cache.CACHE_EVENT_UPDATE_USERS;
-
 public class IssuesRepository extends Repository {
 
     private static final String TAG = IssuesRepository.class.getSimpleName();
@@ -52,44 +39,44 @@ public class IssuesRepository extends Repository {
                 disposables.add(
                         ((EventBroadcaster) asyncDao).getBroadcaster()
                                 .onBackpressureBuffer()
-                                .doOnNext(b -> {
-                                    if (b.equals(CACHE_EVENT_UPDATE_PROJECTS)) {
-                                        Log.i(TAG, "setup: <<<< CACHE_EVENT_UPDATE_PROJECTS");
-                                    }
-                                    if (b.equals(CACHE_EVENT_UPDATE_ISSUES)) {
-                                        Log.i(TAG, "setup: <<<< CACHE_EVENT_UPDATE_ISSUES  ");
-                                    }
-                                    if (b.equals(CACHE_EVENT_UPDATE_ENTRIES)) {
-                                        Log.i(TAG, "setup: <<<< CACHE_EVENT_UPDATE_ENTRIES ");
-                                    }
-                                    if (b.equals(CACHE_EVENT_UPDATE_USERS)) {
-                                        Log.i(TAG, "setup: <<<< CACHE_EVENT_UPDATE_USERS   ");
-                                    }
-                                    if (b.equals(CACHE_EVENT_UPDATE_PROJECT)) {
-                                        Log.i(TAG, "setup: <<<< CACHE_EVENT_UPDATE_PROJECT ");
-                                    }
-                                    if (b.equals(CACHE_EVENT_UPDATE_ISSUE)) {
-                                        Log.i(TAG, "setup: <<<< CACHE_EVENT_UPDATE_ISSUE   ");
-                                    }
-                                    if (b.equals(CACHE_EVENT_UPDATE_ENTRY)) {
-                                        Log.i(TAG, "setup: <<<< CACHE_EVENT_UPDATE_ENTRY   ");
-                                    }
-                                    if (b.equals(CACHE_EVENT_UPDATE_USER)) {
-                                        Log.i(TAG, "setup: <<<< CACHE_EVENT_UPDATE_USER    ");
-                                    }
-                                    if (b.equals(CACHE_EVENT_REMOVE_PROJECT)) {
-                                        Log.i(TAG, "setup: <<<< CACHE_EVENT_REMOVE_PROJECT ");
-                                    }
-                                    if (b.equals(CACHE_EVENT_REMOVE_ISSUE)) {
-                                        Log.i(TAG, "setup: <<<< CACHE_EVENT_REMOVE_ISSUE   ");
-                                    }
-                                    if (b.equals(CACHE_EVENT_REMOVE_ENTRY)) {
-                                        Log.i(TAG, "setup: <<<< CACHE_EVENT_REMOVE_ENTRY   ");
-                                    }
-                                    if (b.equals(CACHE_EVENT_REMOVE_USER)) {
-                                        Log.i(TAG, "setup: <<<< CACHE_EVENT_REMOVE_USER    ");
-                                    }
-                                })
+//                                .doOnNext(b -> {
+//                                    if (b.equals(CACHE_EVENT_UPDATE_PROJECTS)) {
+//                                        Log.i(TAG, "setup: <<<< CACHE_EVENT_UPDATE_PROJECTS");
+//                                    }
+//                                    if (b.equals(CACHE_EVENT_UPDATE_ISSUES)) {
+//                                        Log.i(TAG, "setup: <<<< CACHE_EVENT_UPDATE_ISSUES  ");
+//                                    }
+//                                    if (b.equals(CACHE_EVENT_UPDATE_ENTRIES)) {
+//                                        Log.i(TAG, "setup: <<<< CACHE_EVENT_UPDATE_ENTRIES ");
+//                                    }
+//                                    if (b.equals(CACHE_EVENT_UPDATE_USERS)) {
+//                                        Log.i(TAG, "setup: <<<< CACHE_EVENT_UPDATE_USERS   ");
+//                                    }
+//                                    if (b.equals(CACHE_EVENT_UPDATE_PROJECT)) {
+//                                        Log.i(TAG, "setup: <<<< CACHE_EVENT_UPDATE_PROJECT ");
+//                                    }
+//                                    if (b.equals(CACHE_EVENT_UPDATE_ISSUE)) {
+//                                        Log.i(TAG, "setup: <<<< CACHE_EVENT_UPDATE_ISSUE   ");
+//                                    }
+//                                    if (b.equals(CACHE_EVENT_UPDATE_ENTRY)) {
+//                                        Log.i(TAG, "setup: <<<< CACHE_EVENT_UPDATE_ENTRY   ");
+//                                    }
+//                                    if (b.equals(CACHE_EVENT_UPDATE_USER)) {
+//                                        Log.i(TAG, "setup: <<<< CACHE_EVENT_UPDATE_USER    ");
+//                                    }
+//                                    if (b.equals(CACHE_EVENT_REMOVE_PROJECT)) {
+//                                        Log.i(TAG, "setup: <<<< CACHE_EVENT_REMOVE_PROJECT ");
+//                                    }
+//                                    if (b.equals(CACHE_EVENT_REMOVE_ISSUE)) {
+//                                        Log.i(TAG, "setup: <<<< CACHE_EVENT_REMOVE_ISSUE   ");
+//                                    }
+//                                    if (b.equals(CACHE_EVENT_REMOVE_ENTRY)) {
+//                                        Log.i(TAG, "setup: <<<< CACHE_EVENT_REMOVE_ENTRY   ");
+//                                    }
+//                                    if (b.equals(CACHE_EVENT_REMOVE_USER)) {
+//                                        Log.i(TAG, "setup: <<<< CACHE_EVENT_REMOVE_USER    ");
+//                                    }
+//                                })
 //                                .filter(e -> CACHE_EVENT_UPDATE_ISSUES.equals(e) ||
 //                                        CACHE_EVENT_UPDATE_ISSUE.equals(e) ||
 //                                        CACHE_EVENT_REMOVE_ISSUE.equals(e) ||
