+++ app/src/main/java/org/rares/miner49er/domain/projects/persistence/ProjectsRepository.java
@@ -4,12 +4,10 @@ import android.util.Log;
 import io.reactivex.Flowable;
 import io.reactivex.android.schedulers.AndroidSchedulers;
 import io.reactivex.disposables.CompositeDisposable;
-import io.reactivex.disposables.Disposable;
 import io.reactivex.functions.Consumer;
 import org.rares.miner49er._abstract.Repository;
 import org.rares.miner49er.cache.cacheadapter.AbstractAsyncCacheAdapter;
 import org.rares.miner49er.cache.cacheadapter.InMemoryCacheAdapterFactory;
-import org.rares.miner49er.cache.optimizer.CacheFeeder;
 import org.rares.miner49er.domain.projects.model.ProjectData;
 import org.rares.miner49er.domain.projects.model.ProjectsSort;
 import org.rares.miner49er.persistence.dao.AsyncGenericDao;
@@ -19,19 +17,6 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
-
-import static org.rares.miner49er.cache.Cache.CACHE_EVENT_REMOVE_ENTRY;
-import static org.rares.miner49er.cache.Cache.CACHE_EVENT_REMOVE_ISSUE;
-import static org.rares.miner49er.cache.Cache.CACHE_EVENT_REMOVE_PROJECT;
-import static org.rares.miner49er.cache.Cache.CACHE_EVENT_REMOVE_USER;
-import static org.rares.miner49er.cache.Cache.CACHE_EVENT_UPDATE_ENTRIES;
-import static org.rares.miner49er.cache.Cache.CACHE_EVENT_UPDATE_ENTRY;
-import static org.rares.miner49er.cache.Cache.CACHE_EVENT_UPDATE_ISSUE;
-import static org.rares.miner49er.cache.Cache.CACHE_EVENT_UPDATE_ISSUES;
-import static org.rares.miner49er.cache.Cache.CACHE_EVENT_UPDATE_PROJECT;
-import static org.rares.miner49er.cache.Cache.CACHE_EVENT_UPDATE_PROJECTS;
-import static org.rares.miner49er.cache.Cache.CACHE_EVENT_UPDATE_USER;
-import static org.rares.miner49er.cache.Cache.CACHE_EVENT_UPDATE_USERS;
 
 public class ProjectsRepository extends Repository {
     private static final String TAG = ProjectsRepository.class.getSimpleName();
@@ -50,45 +35,45 @@ public class ProjectsRepository extends Repository {
                 disposables.add(
                         ((EventBroadcaster) asyncDao).getBroadcaster()
                                 .onBackpressureBuffer()
-                                .doOnNext(b -> {
-                                    if (b.equals(CACHE_EVENT_UPDATE_PROJECTS)) {
-                                        Log.v(TAG, "setup: <<<< CACHE_EVENT_UPDATE_PROJECTS");
-                                    }
-                                    if (b.equals(CACHE_EVENT_UPDATE_ISSUES)) {
-                                        Log.v(TAG, "setup: <<<< CACHE_EVENT_UPDATE_ISSUES  ");
-                                    }
-                                    if (b.equals(CACHE_EVENT_UPDATE_ENTRIES)) {
-                                        Log.v(TAG, "setup: <<<< CACHE_EVENT_UPDATE_ENTRIES ");
-                                    }
-                                    if (b.equals(CACHE_EVENT_UPDATE_USERS)) {
-                                        Log.v(TAG, "setup: <<<< CACHE_EVENT_UPDATE_USERS   ");
-                                    }
-                                    if (b.equals(CACHE_EVENT_UPDATE_PROJECT)) {
-                                        Log.v(TAG, "setup: <<<< CACHE_EVENT_UPDATE_PROJECT ");
-                                    }
-                                    if (b.equals(CACHE_EVENT_UPDATE_ISSUE)) {
-                                        Log.v(TAG, "setup: <<<< CACHE_EVENT_UPDATE_ISSUE   ");
-                                    }
-                                    if (b.equals(CACHE_EVENT_UPDATE_ENTRY)) {
-                                        Log.v(TAG, "setup: <<<< CACHE_EVENT_UPDATE_ENTRY   ");
-                                    }
-                                    if (b.equals(CACHE_EVENT_UPDATE_USER)) {
-                                        Log.v(TAG, "setup: <<<< CACHE_EVENT_UPDATE_USER    ");
-                                    }
-                                    if (b.equals(CACHE_EVENT_REMOVE_PROJECT)) {
-                                        Log.v(TAG, "setup: <<<< CACHE_EVENT_REMOVE_PROJECT ");
-                                    }
-                                    if (b.equals(CACHE_EVENT_REMOVE_ISSUE)) {
-                                        Log.v(TAG, "setup: <<<< CACHE_EVENT_REMOVE_ISSUE   ");
-                                    }
-                                    if (b.equals(CACHE_EVENT_REMOVE_ENTRY)) {
-                                        Log.v(TAG, "setup: <<<< CACHE_EVENT_REMOVE_ENTRY   ");
-                                    }
-                                    if (b.equals(CACHE_EVENT_REMOVE_USER)) {
-                                        Log.v(TAG, "setup: <<<< CACHE_EVENT_REMOVE_USER    ");
-                                    }
-                                    Log.i(TAG, "setup: " + asyncDao.getAll(true).blockingGet().size());
-                                })
+//                                .doOnNext(b -> {
+//                                    if (b.equals(CACHE_EVENT_UPDATE_PROJECTS)) {
+//                                        Log.v(TAG, "setup: <<<< CACHE_EVENT_UPDATE_PROJECTS");
+//                                    }
+//                                    if (b.equals(CACHE_EVENT_UPDATE_ISSUES)) {
+//                                        Log.v(TAG, "setup: <<<< CACHE_EVENT_UPDATE_ISSUES  ");
+//                                    }
+//                                    if (b.equals(CACHE_EVENT_UPDATE_ENTRIES)) {
+//                                        Log.v(TAG, "setup: <<<< CACHE_EVENT_UPDATE_ENTRIES ");
+//                                    }
+//                                    if (b.equals(CACHE_EVENT_UPDATE_USERS)) {
+//                                        Log.v(TAG, "setup: <<<< CACHE_EVENT_UPDATE_USERS   ");
+//                                    }
+//                                    if (b.equals(CACHE_EVENT_UPDATE_PROJECT)) {
+//                                        Log.v(TAG, "setup: <<<< CACHE_EVENT_UPDATE_PROJECT ");
+//                                    }
+//                                    if (b.equals(CACHE_EVENT_UPDATE_ISSUE)) {
+//                                        Log.v(TAG, "setup: <<<< CACHE_EVENT_UPDATE_ISSUE   ");
+//                                    }
+//                                    if (b.equals(CACHE_EVENT_UPDATE_ENTRY)) {
+//                                        Log.v(TAG, "setup: <<<< CACHE_EVENT_UPDATE_ENTRY   ");
+//                                    }
+//                                    if (b.equals(CACHE_EVENT_UPDATE_USER)) {
+//                                        Log.v(TAG, "setup: <<<< CACHE_EVENT_UPDATE_USER    ");
+//                                    }
+//                                    if (b.equals(CACHE_EVENT_REMOVE_PROJECT)) {
+//                                        Log.v(TAG, "setup: <<<< CACHE_EVENT_REMOVE_PROJECT ");
+//                                    }
+//                                    if (b.equals(CACHE_EVENT_REMOVE_ISSUE)) {
+//                                        Log.v(TAG, "setup: <<<< CACHE_EVENT_REMOVE_ISSUE   ");
+//                                    }
+//                                    if (b.equals(CACHE_EVENT_REMOVE_ENTRY)) {
+//                                        Log.v(TAG, "setup: <<<< CACHE_EVENT_REMOVE_ENTRY   ");
+//                                    }
+//                                    if (b.equals(CACHE_EVENT_REMOVE_USER)) {
+//                                        Log.v(TAG, "setup: <<<< CACHE_EVENT_REMOVE_USER    ");
+//                                    }
+//                                    Log.i(TAG, "setup: " + asyncDao.getAll(true).blockingGet().size());
+//                                })
 //                                .throttleLast(500, TimeUnit.MILLISECONDS)
                                 .subscribe(o -> refreshData(true)));
             }
@@ -121,34 +106,17 @@ public class ProjectsRepository extends Repository {
         disposables.add(
 //                Flowable.merge(
                 userActionsObservable
-                        .map(action -> getData())       // todo: [getFromCache|getFromDB]|getFromNetwork -> take 1 ->
-                        .startWith(() -> {
-                            Disposable cacheFillDisposable = cacheFeeder.enqueueCacheFill();
-                            if (cacheFillDisposable != null) {
-                                disposables.add(cacheFillDisposable);
-                                return Collections.<List<ProjectData>>emptyList().iterator();
-                            } else {
-                                return Collections.singleton(getData()).iterator();
-                            }
-                        })
-
-//                )
                         .onBackpressureBuffer()
-                        // FIXME: 3/1/19 | comment next line out, do not use throttle in ViewModelCache events and fix the LayoutManager!
-//                        .debounce(50, TimeUnit.MILLISECONDS)
                         .onErrorResumeNext(Flowable.fromIterable(Collections.emptyList()))
-//                        .doOnNext(x -> Log.i(TAG, "registerSubscriber: -> " + x.size()))
-                        .doOnError((e) -> Log.e(TAG, "registerSubscriber: ", e))
-//                        .doOnComplete(() -> Log.d(TAG, "registerSubscriber: success"))
-//                        .doOnSubscribe(x -> Log.i(TAG, "registerSubscriber: on subscribe"))
+                        .map(action -> getData())       // todo: [getFromCache|getFromDB]|getFromNetwork -> take 1 ->
                         .observeOn(AndroidSchedulers.mainThread())
                         .subscribe(consumer));
     }
 
     private List<ProjectData> getData() {
         List<ProjectData> toReturn = asyncDao.getAll(true)
-                .timeout(250, TimeUnit.MILLISECONDS)   // 1 second should be enough to get data (from cache)
-                .doOnError(e -> Log.e(TAG, "getData: timeout? ", e))
+                .timeout(250, TimeUnit.MILLISECONDS)   // should be enough to get data (from cache)
+                .doOnError(e -> Log.e(TAG, "getData: setup timeout? ", e))
                 .onErrorReturnItem(Collections.emptyList())
                 // something gets stuck when there is no data in the db
                 // and data is being transferred
@@ -161,6 +129,4 @@ public class ProjectsRepository extends Repository {
         }
         return clones;
     }
-
-    private final CacheFeeder cacheFeeder = new CacheFeeder();
 }
